refactor: Use `u16` for `start_idx` in `subscribe_to_next_responses`

+++ crates/db-postgres/src/postgres_dao.rs
@@ -2320,7 +2320,7 @@ async fn delay_response(
 async fn get_responses_with_offset(
     tx: &Transaction<'_>,
     execution_id: &ExecutionId,
-    skip_rows: usize,
+    skip_rows: u16,
 ) -> Result<Vec<JoinSetResponseEventOuter>, DbErrorRead> {
     let rows = tx
             .query(
@@ -2338,7 +2338,7 @@ async fn get_responses_with_offset(
                  OFFSET $2",
                  &[
                      &execution_id.to_string(),
-                     &(skip_rows as i64),
+                     &(i64::from(skip_rows)),
                  ]
             )
             .await
@@ -3048,7 +3048,7 @@ impl DbConnection for PostgresConnection {
     async fn subscribe_to_next_responses(
         &self,
         execution_id: &ExecutionId,
-        start_idx: usize,
+        start_idx: u16,
         timeout_fut: Pin<Box<dyn Future<Output = ()> + Send>>,
     ) -> Result<Vec<JoinSetResponseEventOuter>, DbErrorReadWithTimeout> {
         debug!("next_responses");
